+++ Assets/Scripts/ProjectStatusJson.cs
@@ -1,6 +1,6 @@
 [System.Serializable]
 public class ProjectStatusJson
 {
-    public float totalStaked;
+    public double totalStaked;
     public float apy;
 }
